feat(agent-interface): provide agent interface
